core: move from src/pages to src/app (NextJS App Router way)
