+++ package.json
@@ -18,6 +18,7 @@
     "@types/react": "18.2.20",
     "@types/react-dom": "18.2.7",
     "autoprefixer": "10.4.14",
+    "classnames": "^2.3.2",
     "date-fns": "^2.30.0",
     "next": "13.4.13",
     "next-sanity": "^5.2.6",
