--- src/pages/_document.tsx
@@ -1,17 +1,0 @@
-import { Html, Head, Main, NextScript } from 'next/document'
-
-export default function Document() {
-  return (
-    <Html lang="en">
-      <Head>
-        <link rel="preconnect" href="https://fonts.googleapis.com" />
-        <link rel="preconnect" href="https://fonts.gstatic.com" crossOrigin={""} />
-        <link href="https://fonts.googleapis.com/css2?family=Montserrat:wght@400;700&family=Noto+Sans+JP:wght@100;400&display=swap" rel="stylesheet"/ >
-      </Head>
-      <body className="bg-black">
-        <Main />
-        <NextScript />
-      </body>
-    </Html>
-  )
-}
