--- src/types.ts
@@ -1,16 +1,0 @@
-export type EventType = {
-  title: string;
-  bgColor: string;
-  image: any;
-}
-
-export type CommunityEvent = {
-  eventType: EventType;
-  title: string;
-  mergedTitle: string;
-  startDate: string;
-  endDate: string;
-  url: string;
-  bgColor: string;
-  image: any;
-}
